Configure dev environment
