+++ .npmignore
@@ -1,0 +1,3 @@
+/*
+!/lib/*.js
+!index.js
