+++ index.js
@@ -1,1 +1,1 @@
-module.exports = require('./helpers')
+module.exports = require('./lib/helpers')
