+++ package.json
@@ -4,7 +4,22 @@
   "description": "Helper functions to simplify Jest integration with Lighthouse",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "npm run build -- --watch",
+    "build": "tsc",
+    "format": "prettier --write \"*.{js,md}\" \"src/**/*.js\"",
+    "lint": "tsc --noEmit && eslint . --ext .js,.ts",
+    "prepublishOnly": "npm run build"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
+    }
+  },
+  "lint-staged": {
+    "**/*.{js,ts,md}": [
+      "prettier --write"
+    ],
+    "*.{js,ts}": "eslint --fix"
   },
   "repository": {
     "type": "git",
@@ -23,5 +38,18 @@
   "homepage": "https://github.com/mmarkelov/jest-lighthouse#readme",
   "dependencies": {
     "lighthouse": "^7.0.1"
+  },
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^4.14.1",
+    "@typescript-eslint/parser": "^4.14.1",
+    "eslint": "^7.19.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-config-prettier": "^7.2.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-prettier": "^3.3.1",
+    "husky": "^4.3.8",
+    "lint-staged": "^10.5.3",
+    "prettier": "^2.2.1",
+    "typescript": "^4.1.3"
   }
 }
